JBIDE-14794 - install from central
https://issues.jboss.org/browse/JBIDE-14794
install from central shows "selection has been modified" as if p2 decided
to install something different and p2 remediation not used/shown
