+++ examples/plugins/org.jboss.tools.project.examples/META-INF/MANIFEST.MF
@@ -45,7 +45,10 @@ Require-Bundle: org.eclipse.ui,
  org.jboss.tools.foundation.core;bundle-version="1.0.0",
  org.jboss.tools.foundation.ui,
  org.eclipse.ui.browser;bundle-version="3.4.0",
- org.jboss.tools.as.runtimes.integration;bundle-version="2.4.100"
+ org.jboss.tools.as.runtimes.integration;bundle-version="2.4.100",
+ org.eclipse.equinox.p2.repository,
+ org.eclipse.equinox.p2.director;bundle-version="2.3.0",
+ org.eclipse.equinox.p2.metadata.repository;bundle-version="1.2.100"
 Bundle-ActivationPolicy: lazy
 Bundle-Localization: plugin
 Export-Package: org.jboss.tools.project.examples,
